+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## NEXT RELEASE
  - add link to the event card on Photo's activity list - #799
  - add feature to show Past & Future events on Activity canonical page - #806 #804
+ - add standard (numeric) pager on Activity canonical page listing of past events - #805
 
 ## 1.6.1 (2021-04-23)
  - update drupal/core (8.9.13 => 8.9.14)